Create first draft of BoardScreen (#16)
* client: create isJoined field in Redux global store

isJoined representes whether the player should be presented with one of
the two "lobby" screens, or the screen with the actual game board. Have
they "joined" a game and started playing, or are they still waiting to
either set their display name, or for the game to begin?

* client: create BoardScreen and a supplementary Board component

* client: store board in global Redux store, and display it with Cards

another god commit. this project's git history is gonna be looking a bit
hairy

* client: add :hover and :active styles to the Card component

* client: prototype/stub a rough draft of the BoardScreen's UI

* client: correct typos in comments